manager/pyproject.toml
[tool.poetry]
name = "leannlsearch"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.31.0"
weaviate-client = "^3.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
